# Week11_G4_API/Models/Course.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;

namespace Week11_G4_API.Models
{
    public partial class Course
    {
        public string CourseId { get; set; }
        public string CourseName { get; set; }
        public DateTime? CourseStartDate { get; set; }
        public DateTime? CourseEndDate { get; set; }
        public string InstructorId { get; set; }
        public string CreditId { get; set; }
    }
}