# Week11_G4_API/Models/Credit.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;

namespace Week11_G4_API.Models
{
    public partial class Credit
    {
        public string CreditId { get; set; }
        public int? CreditValue { get; set; }
    }
}