# Week11_G4_API/Models/Instructor.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;

namespace Week11_G4_API.Models
{
    public partial class Instructor
    {
        public string InstructorId { get; set; }
        public string InstructorFirstName { get; set; }
        public string InstructorLastName { get; set; }
        public string InstructorEmail { get; set; }
        public decimal? InstructorSalary { get; set; }
    }
}